# docs/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: team@testseer.io
    name: TestSeer Team
  description: "TestSeer analyses Java service codebases to produce a structured knowledge\
    \ graph of classes, endpoints, and outbound HTTP calls. Use the Query API to ask\
    \ questions about your services — reachability, impact analysis, shared type consumers\
    \ — and always get freshness metadata alongside results."
  license:
    name: Proprietary
  title: TestSeer API
  version: 0.1.0
servers:
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /v1/facts/by-file:
    get:
      description: "Returns all indexed symbols (CLASS, ENDPOINT) for the given file\
        \ paths within a service. Used by the MCP server to map GitHub PR changed\
        \ files to their indexed endpoints and classes."
      operationId: getByFile
      parameters:
      - description: Service identifier
        in: query
        name: serviceId
        required: true
        schema:
          type: string
      - description: File paths to look up
        in: query
        name: filePaths
        required: false
        schema:
          type: array
          items:
            type: string
      - description: Organisation ID
        in: query
        name: orgId
        required: false
        schema:
          type: string
          default: acme
      - description: Repository name
        in: query
        name: repo
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListSymbolFactView"
          description: Symbols found (may be empty if files not indexed)
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListSymbolFactView"
          description: Service not indexed
      summary: Get symbol facts for specific file paths
      tags:
      - Query — Facts
  /v1/facts/class:
    get:
      description: "Returns all indexed symbol facts for the given class FQN within\
        \ a service. Results are Redis-cached. Returns 404 if the service has never\
        \ been indexed, 202 if indexing is currently in progress."
      operationId: getClassFacts
      parameters:
      - description: Service identifier
        in: query
        name: serviceId
        required: true
        schema:
          type: string
      - description: "Fully-qualified class name, e.g. com.example.OrdersController"
        in: query
        name: symbolFqn
        required: true
        schema:
          type: string
      - description: Organisation ID
        in: query
        name: orgId
        required: false
        schema:
          type: string
          default: acme
      - description: Repository name
        in: query
        name: repo
        required: false
        schema:
          type: string
          default: ""
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListSymbolFactView"
          description: Facts returned
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListSymbolFactView"
          description: Indexing in progress — stale data may be returned
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListSymbolFactView"
          description: Service not indexed
      summary: Get symbol facts for a class
      tags:
      - Query — Facts
...
  /v1/facts/external-endpoints:
    get:
      description: "Returns config-resolved external and internal partner HTTP targets\
        \ (e.g. Hyvee LMS notify URL, OIS partnerpublishingdetails). Filter by env\
        \ lane, partner slug, or flow step."
      operationId: getExternalEndpoints
      parameters:
      - in: query
        name: serviceId
        required: true
        schema:
          type: string
      - in: query
        name: orgId
        required: false
        schema:
          type: string
          default: acme
      - in: query
        name: repo
        required: false
        schema:
          type: string
          default: ""
      - in: query
        name: env
        required: false
        schema:
          type: string
      - in: query
        name: partner
        required: false
        schema:
          type: string
      - in: query
        name: flowStep
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListExternalEndpointFactView"
          description: External endpoint facts returned
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListExternalEndpointFactView"
          description: Indexing in progress
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListExternalEndpointFactView"
          description: Service not indexed
      summary: Get external/partner HTTP endpoint facts for a service
      tags:
      - Query — Facts
...
  /v1/facts/outbound:
    get:
      description: "Returns all detected outbound HTTP calls made by the service.\
        \ Includes RestClient/WebClient chain calls, RestTemplate explicit methods,\
        \ and FeignClient interface declarations. `httpMethod` and `path` are null\
        \ when only client field presence was detected (e.g. a dynamic URI in a variable).\
        \ Optionally filtered to a single caller class."
      operationId: getOutboundFacts
      parameters:
      - description: Service identifier
        in: query
        name: serviceId
        required: true
        schema:
          type: string
      - description: Organisation ID
        in: query
        name: orgId
        required: false
        schema:
          type: string
          default: acme
      - description: Repository name
        in: query
        name: repo
        required: false
        schema:
          type: string
          default: ""
      - description: Filter by caller class FQN (optional)
        in: query
        name: sourceSymbol
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListOutboundCallView"
          description: Outbound call facts returned
        "202":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListOutboundCallView"
          description: Indexing in progress
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResponseEnvelopeListOutboundCallView"
          description: Service not indexed
      summary: Get outbound HTTP call facts for a service
      tags:
      - Query — Facts
...
components:
  schemas:
...
    ExternalEndpointFactView:
      type: object
      properties:
        authScheme:
          type: string
        boundary:
          type: string
        callerClassFqn:
          type: string
        clientClassFqn:
          type: string
        confidence:
          type: number
          format: double
        configKey:
          type: string
        endpointId:
          type: string
        envLane:
          type: string
        evidenceSource:
          type: string
        flowStep:
          type: string
        httpMethod:
          type: string
        indexedAt:
          type: string
          format: date-time
        operation:
          type: string
        partnerSlug:
          type: string
        urlResolved:
          type: string
        urlTemplate:
          type: string
...
    OutboundCallView:
      type: object
      properties:
        confidence:
          type: number
          format: double
        evidenceSource:
          type: string
        httpMethod:
          type: string
        indexedAt:
          type: string
          format: date-time
        path:
          type: string
        sourceSymbol:
          type: string
...
    ResponseEnvelopeListExternalEndpointFactView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ExternalEndpointFactView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListOutboundCallView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/OutboundCallView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    ResponseEnvelopeListSymbolFactView:
      type: object
      properties:
        commitSha:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SymbolFactView"
        freshnessStatus:
          type: string
          enum:
          - CURRENT
          - STALE
          - INDEXING
          - NOT_INDEXED
        indexedAt:
          type: string
          format: date-time
        liveConfigEnv:
          type: string
        liveConfigStatus:
          type: string
        livePubSubSkippedCount:
          type: integer
          format: int32
        livePubSubStatus:
          type: string
        livePubSubVerifiedCount:
          type: integer
          format: int32
        rulePackHash:
          type: string
        schemaVersion:
          type: string
...
    SymbolFactView:
      type: object
      properties:
        attributes:
          type: string
        confidence:
          type: number
          format: double
        evidenceSource:
          type: string
        indexedAt:
          type: string
          format: date-time
        symbolFqn:
          type: string
        symbolKind:
          type: string
...